requirements.txt
# Core Dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic-ai>=0.0.14

# Flask Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
werkzeug>=3.0.0

# LangGraph & LangChain
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pgvector>=0.2.0

# API Clients
openai>=1.10.0
requests>=2.31.0
httpx>=0.26.0
google-auth>=2.22.0

# Password Hashing
argon2-cffi>=23.1.0

# Templates
Jinja2>=3.1.0

# Configuration
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
ruff>=0.1.9
